RAC-607: move mocked dependencies to src to allow micro frontend to use it
