+++ front-packages/shared/src/tests/index.ts
@@ -1,0 +1,2 @@
+export * from './mockedDependencies';
+export * from './utils';
